==> code/integrative/meqtl/04_me-qtl_analysis/04_02_most_likely_genotype/05_cell_type_specific_analysis/03_rslt_analysis_groups.Rmd <==
---
title: "Cell-type specific enrichment analysis"
subtitle: "Dexamethasone-Stimulated Human Array Project"
output:
  html_notebook
editor_options:
  chunk_output_type: inline
---

```{r, warning=FALSE, message=FALSE}
library(data.table)
library(dplyr)
library(ggplot2)
library(corrplot)
library(factoextra)
library(viridis)
```

```{r}
cbPalette <- c( "#0072B2", "#009E73", "#E69F00", "#F0E442", "#D55E00", "#CC79A7", "#56B4E9", "#999999")
```

Load groups

```{r}
opposite.fc.cpg.ids     <- fread("~/bio/code/mpip/dex-stim-human-array/output/data/integrative/matrixEQTL/meqtls/meqtl_parallel_and_opposite_fc_groups/meqtl_opposite_fc_gr_cpg_list.csv")
parallel.fc.delta.ids   <- fread("~/bio/code/mpip/dex-stim-human-array/output/data/integrative/matrixEQTL/meqtls/meqtl_parallel_and_opposite_fc_groups/meqtl_parallel_fc_gr_delta_cpg_lst.csv")
parallel.fc.veh.dex.ids <- fread("~/bio/code/mpip/dex-stim-human-array/output/data/integrative/matrixEQTL/meqtls/meqtl_parallel_and_opposite_fc_groups/meqtl_parallel_fc_gr_veh_dex_cpg_lst.csv")
```

```{r}
treatment   <- "veh"

pval.veh.df <- fread(paste0("~/bio/code/mpip/dex-stim-human-array/output/data/integrative/cell_type_enrichment/dnam_cell_type_enrichment_pvals_",treatment, ".csv"))

pval.veh.bcc.df <- pval.veh.df[, 1:13]
```

```{r}
fdr.bcc.df <- matrix(p.adjust(as.vector(as.matrix(pval.veh.bcc.df[, 2:13])), method='fdr'), 
                     ncol=12) %>%
  data.frame()

fdr.bcc.df <- cbind(pval.veh.bcc.df$CpG_ID, fdr.bcc.df)
colnames(fdr.bcc.df) <- colnames(pval.veh.bcc.df)
```

```{r}
sign.pval.df <- fdr.bcc.df %>% reshape2::melt(measure.vars = colnames(fdr.bcc.df)[2:13]) %>% setDT()
colnames(sign.pval.df) <- c("CpG_ID", "Type", "fdr")
sign.pval.df <- sign.pval.df[fdr <= 0.05]

sign.pval.df[CpG_ID %in% opposite.fc.cpg.ids$CpG_ID, Model := "opposite_fc"]
sign.pval.df[CpG_ID %in% parallel.fc.delta.ids$CpG_ID, Model := "parallel_fc_delta"]
sign.pval.df[CpG_ID %in% parallel.fc.veh.dex.ids$CpG_ID, Model := "parallel_fc_veh_dex"] 

sign.pval.df <- na.omit(sign.pval.df)
sign.pval.df
```

```{r}
veh.sign.pval.df <- sign.pval.df
veh.sign.pval.df[["Treatment"]] <- treatment
```

```{r include = F}
ggplot(sign.pval.df, aes(x = Type)) +
  geom_bar(stat = "count", alpha = 1, position = position_dodge(width = .9), fill = cbPalette[3]) +
  # geom_text(aes(label =  scales::percent(prop.table(stat(count)))), 
  # geom_text(aes(label = scales::percent(stat(count) / length(unique(sign.pval.df$CpG_ID)))), 
  geom_text(aes(label = scales::percent(stat(count) / nrow(fdr.bcc.df))), 
            stat = "count", vjust = -0.5, size = 3, position = position_dodge(width = .9)) + 
  theme(legend.position = "bottom", # c(0.9, 0.9), 
        legend.title = element_blank(), 
        panel.grid.major = element_blank(), 
        panel.background = element_blank(),
        plot.title = element_text(size = 10), 
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 8)) + 
  labs(title = "Distribution of CpGs significant at FDR = 0.05 across 12 blood cell types",
       x = "Blood cell type", y = "")
```

```{r fig.width = 7, fig.height = 8}
sign.pval.df[Model == "opposite_fc", CpG_ID] %>% unique() %>% length()
sign.pval.df[Model == "parallel_fc_delta", CpG_ID] %>% unique() %>% length()
sign.pval.df[Model == "parallel_fc_veh_dex", CpG_ID] %>% unique() %>% length()

stats <- sign.pval.df %>% group_by(Model, Type) %>% count() %>% data.frame() %>% setDT()
stats[Model == "opposite_fc", percent := scales::percent(n / nrow(opposite.fc.cpg.ids))]
stats[Model == "parallel_fc_delta", percent := scales::percent(n / nrow(parallel.fc.delta.ids))]
stats[Model == "parallel_fc_veh_dex", percent := scales::percent(n / nrow(parallel.fc.veh.dex.ids))]

stats %>% 
  ggplot(aes(x = Type, y = n, fill = Model)) +
  geom_bar(stat = "identity", alpha = 1, position = "dodge") +
  facet_wrap(~ Model, nrow = 3, scales = "free") +
  geom_text(aes(label = percent), vjust = -0.5, size = 5, position = position_dodge(width = .9)) +
  theme(legend.position = "bottom", # c(0.9, 0.9), 
        legend.title = element_blank(), 
        panel.grid.major = element_blank(), 
        panel.background = element_blank(),
        plot.title = element_text(size = 12), 
        axis.title = element_text(size = 12),
        axis.text.x = element_text(size = 12),
        axis.text.y = element_text(size = 12),
        strip.text = element_text(size=12)) + 
  labs(title = "Distribution of CpGs significant at FDR = 0.05 across 12 blood cell types",
       x = "Blood cell type", y = "") +
  scale_fill_brewer(palette = "Dark2")
```

```{r}
ggplot(sign.pval.df, aes(x = Type, group = Model, fill = Model)) + 
  # geom_density(aes(y = stat(count) / nrow(fdr.bcc.df), color = Model),
  geom_density(aes(y = ..prop.., color = Model), 
           stat = "count",
           alpha = 0.2) + 
  scale_y_continuous(labels = scales::percent) + 
  labs(x = "Blood cell type", y = "Frequency", 
       title = "Distribution of CpGs significant at FDR = 0.05 across 12 blood cell types") + 
  theme(legend.title = element_blank(), 
        legend.position = "bottom", 
        panel.grid.major = element_blank(), 
        panel.background = element_blank(), 
        plot.title = element_text(size = 8), 
        axis.title = element_text(size = 8), 
        axis.text.x = element_text(angle = 0, hjust = 0.5)) +
   scale_fill_brewer(palette = "Dark2") +
   scale_color_brewer(palette = "Dark2")
```

Baseline TCA models

```{r}
tca.mu.df <- fread(paste0("~/bio/code/mpip/dex-stim-human-array/output/data/integrative/cell_type_enrichment/1_tca/dnam_cell_type_tca_mu_chr_all_", treatment, ".csv"))
```

```{r}
tca.mu.sign.df <- tca.mu.df[rn %in% sign.pval.df$CpG_ID]
tca.mu.sign.df <- tca.mu.sign.df %>% reshape2::melt() 
colnames(tca.mu.sign.df) <- c("CpG_ID", "Type", "tca_mu_val")

tca.mu.sign.df$Type <- sub("salas.", "", tca.mu.sign.df$Type)
tca.sign.mu.pval.df <- left_join(sign.pval.df, tca.mu.sign.df)
```

```{r}
tca.sign.mu.median.veh.df <- aggregate(tca_mu_val ~ Model + Type, tca.sign.mu.pval.df, median)
tca.sign.mu.median.veh.df[["Treatment"]] <- treatment
```

DEX

```{r}
treatment   <- "dex"

pval.dex.df <- fread(paste0("~/bio/code/mpip/dex-stim-human-array/output/data/integrative/cell_type_enrichment/dnam_cell_type_enrichment_pvals_",treatment, ".csv"))

pval.dex.bcc.df <- pval.dex.df[, 1:13]
```

```{r}
fdr.bcc.df <- matrix(p.adjust(as.vector(as.matrix(pval.dex.bcc.df[, 2:13])), method ='fdr'), 
                     ncol=12) %>%
  data.frame()

fdr.bcc.df <- cbind(pval.dex.bcc.df$CpG_ID, fdr.bcc.df)
colnames(fdr.bcc.df) <- colnames(pval.dex.bcc.df)
```

```{r}
sign.pval.df <- fdr.bcc.df %>% reshape2::melt(measure.vars = colnames(fdr.bcc.df)[2:13]) %>% setDT()
colnames(sign.pval.df) <- c("CpG_ID", "Type", "fdr")
sign.pval.df <- sign.pval.df[fdr <= 0.05]

sign.pval.df[CpG_ID %in% opposite.fc.cpg.ids$CpG_ID, Model := "opposite_fc"]
sign.pval.df[CpG_ID %in% parallel.fc.delta.ids$CpG_ID, Model := "parallel_fc_delta"]
sign.pval.df[CpG_ID %in% parallel.fc.veh.dex.ids$CpG_ID, Model := "parallel_fc_veh_dex"] 

sign.pval.df <- na.omit(sign.pval.df)
sign.pval.df
```

```{r}
dex.sign.pval.df <- sign.pval.df
dex.sign.pval.df[["Treatment"]] <- treatment
```

```{r fig.width = 7, fig.height = 8}
sign.pval.df[Model == "opposite_fc", CpG_ID] %>% unique() %>% length()
sign.pval.df[Model == "parallel_fc_delta", CpG_ID] %>% unique() %>% length()
sign.pval.df[Model == "parallel_fc_veh_dex", CpG_ID] %>% unique() %>% length()

stats <- sign.pval.df %>% group_by(Model, Type) %>% count() %>% data.frame() %>% setDT()
stats[Model == "opposite_fc", percent := scales::percent(n / nrow(opposite.fc.cpg.ids))]
stats[Model == "parallel_fc_delta", percent := scales::percent(n / nrow(parallel.fc.delta.ids))]
stats[Model == "parallel_fc_veh_dex", percent := scales::percent(n / nrow(parallel.fc.veh.dex.ids))]

stats %>% 
  ggplot(aes(x = Type, y = n, fill = Model)) +
  geom_bar(stat = "identity", alpha = 1, position = "dodge") +
  facet_wrap(~ Model, nrow = 3, scales = "free") +
  geom_text(aes(label = percent), vjust = -0.5, size = 5, position = position_dodge(width = .9)) +
  theme(legend.position = "bottom", # c(0.9, 0.9), 
        legend.title = element_blank(), 
        panel.grid.major = element_blank(), 
        panel.background = element_blank(),
        plot.title = element_text(size = 12), 
        axis.title = element_text(size = 12),
        axis.text.x = element_text(size = 12),
        axis.text.y = element_text(size = 12),
        strip.text = element_text(size=12)) + 
  labs(title = "Distribution of CpGs significant at FDR = 0.05 across 12 blood cell types",
       x = "Blood cell type", y = "") +
  scale_fill_brewer(palette = "Dark2")
```

```{r}
ggplot(sign.pval.df, aes(x = Type, group = Model, fill = Model)) + 
  geom_density(aes(y = ..prop.., color = Model), 
           stat = "count",
           alpha = 0.2) + 
  scale_y_continuous(labels = scales::percent) + 
  labs(x = "Blood cell type", y = "Frequency", 
       title = "Distribution of CpGs significant at FDR = 0.05 across 12 blood cell types") + 
  theme(legend.title = element_blank(), 
        legend.position = "bottom", 
        panel.grid.major = element_blank(), 
        panel.background = element_blank(), 
        plot.title = element_text(size = 8), 
        axis.title = element_text(size = 8), 
        axis.text.x = element_text(angle = 0, hjust = 0.5)) +
   scale_fill_brewer(palette = "Dark2") +
   scale_color_brewer(palette = "Dark2")

```

Read TCA dex models

```{r}
tca.mu.df <- fread(paste0("~/bio/code/mpip/dex-stim-human-array/output/data/integrative/cell_type_enrichment/1_tca/dnam_cell_type_tca_mu_chr_all_", treatment, ".csv"))
```

```{r}
tca.mu.sign.df <- tca.mu.df[rn %in% sign.pval.df$CpG_ID]
tca.mu.sign.df <- tca.mu.sign.df %>% reshape2::melt() 
colnames(tca.mu.sign.df) <- c("CpG_ID", "Type", "tca_mu_val")

tca.mu.sign.df$Type <- sub("salas.", "", tca.mu.sign.df$Type)
tca.sign.mu.pval.df <- left_join(sign.pval.df, tca.mu.sign.df)
```

```{r}
tca.sign.mu.median.dex.df <- aggregate(tca_mu_val ~ Model + Type, tca.sign.mu.pval.df, median)
tca.sign.mu.median.dex.df[["Treatment"]] <- treatment
```

```{r fig.width = 7, fig.height = 7}
tca.sign.mu.median.df <- rbind(tca.sign.mu.median.veh.df, tca.sign.mu.median.dex.df)

cbPalette <- c( "#0072B2", "#009E73", "#E69F00", "#F0E442", "#D55E00", "#CC79A7", "#56B4E9", "#999999")
ggplot(tca.sign.mu.median.df, aes(y = tca_mu_val, x = Type, fill = Treatment)) +
  facet_wrap(~ Model, nrow = 3) +
  geom_bar(stat = "identity", position = position_dodge()) +
  scale_fill_viridis(discrete = TRUE, alpha = 0.7, option = "D") +
  labs(title = " ", y = "Median of all estimates for the mean of each blood cell type in each CpG", x = " ") + 
  theme(legend.position = "bottom", 
        legend.title = element_blank(),
        panel.grid.major = element_blank(),
        panel.background = element_blank(),
        plot.title = element_text(size = 10),
        axis.title = element_text(size = 8),
        axis.text.x = element_text(angle = 20, hjust = 0.5))+
  scale_fill_manual(values = cbPalette[2:3]) 
```

## Distribution plots for Basleine and Dex together

```{r}
sign.pval.df <- rbind(veh.sign.pval.df, dex.sign.pval.df)
```

```{r}
# perc.labels <- (table(sign.pval.df$Type) / nrow(fdr.bcc.df)) %>% as.numeric() %>% scales::percent()

ggplot(sign.pval.df, aes(x = Type, fill = Treatment)) +
  geom_bar(stat = "count", alpha = 1, position = position_dodge(width = .9)) +
  # geom_text(aes(label =  scales::percent(prop.table(stat(count)))), 
  # geom_text(aes(label = scales::percent(stat(count) / length(unique(sign.pval.df$CpG_ID)))), 
  geom_text(aes(label = scales::percent(stat(count) / nrow(fdr.bcc.df), accuracy = 1)), 
            stat = "count", vjust = -0.5, size = 3, position = position_dodge(width = .9)) + 
  theme(legend.position = "bottom", # c(0.9, 0.9), 
        legend.title = element_blank(), 
        panel.grid.major = element_blank(), 
        panel.background = element_blank(),
        plot.title = element_text(size = 10), 
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 10)) + 
  labs(title = "Distribution of CpGs significant at FDR = 0.05 across 12 blood cell types",
       x = "", y = "") +
  scale_fill_manual(values = cbPalette[2:3]) 
```

```{r}
ggplot(sign.pval.df, aes(x = Type, group = Treatment, fill = Treatment)) + 
  geom_density(aes(y = stat(count) / nrow(fdr.bcc.df), color = Treatment), 
           stat = "count",
           alpha = 0.2) + 
  scale_y_continuous(labels = scales::percent) + 
  labs(x = " ", y = "", 
       title = "Distribution of CpGs significant at FDR = 0.05 across 12 blood cell types") + 
  theme(legend.title = element_blank(), 
        legend.position = "bottom", 
        panel.grid.major = element_blank(), 
        panel.background = element_blank(), 
        plot.title = element_text(size = 8), 
        axis.title = element_text(size = 8), 
        axis.text.x = element_text(angle = 0, hjust = 0.5)) +
  scale_color_manual(values = cbPalette[2:3]) +
  scale_fill_manual(values = cbPalette[2:3]) 
```

# Cell-type specificity on GR-induced (delta)-meQTLs

```{r}
ind.meqtl.delta.df      <- fread("~/bio/code/mpip/dex-stim-human-array/output/data/integrative/matrixEQTL/meqtls/me-qtl_cis_result_delta_fdr_005.csv") 
delta.meqtl.cpgs        <- ind.meqtl.delta.df$gene %>% unique() 
```

## Check if changes are significant

```{r}
sign.pval.delta.mqtl.df <- sign.pval.df[CpG_ID %in% delta.meqtl.cpgs, ] 
blood.cell.types        <- levels(sign.pval.delta.mqtl.df$Type)

test.res.df <- lapply(blood.cell.types, function(i){  # for each blood cell type
  dex.fdr.lst <- sign.pval.delta.mqtl.df[Type == i][Treatment == "dex", fdr]
  veh.fdr.lst <- sign.pval.delta.mqtl.df[Type == i][Treatment == "veh", fdr]
  test.rslt   <- var.test(dex.fdr.lst, veh.fdr.lst)
 
  return(data.frame(Type = i, "p-value" = test.rslt$p.value))
  }) %>%
  bind_rows()
```

```{r}
sign.pval.delta.mqtl.df <- left_join(sign.pval.delta.mqtl.df, test.res.df)
ggplot(sign.pval.delta.mqtl.df, aes(x = Type, fill = Treatment)) +
  geom_bar(stat = "count", alpha = 1, position = position_dodge(width = .9)) +
  geom_text(aes(label = scales::percent(stat(count) / length(delta.meqtl.cpgs), accuracy = 1)), 
            stat = "count", vjust = 1.5, size = 5, position = position_dodge(width = .9)) + 
  geom_text(data = subset(sign.pval.delta.mqtl.df[Treatment == "veh"], p.value <= 0.05),
            aes(label = "*"), stat = "count", vjust = .5,  fill = cbPalette[8], size = 16) +
  theme(legend.position = "bottom", # c(0.9, 0.9), 
        legend.title = element_blank(), 
        legend.text = element_text(size = 18),
        panel.grid.major = element_blank(), 
        panel.background = element_blank(),
        plot.title = element_text(size = 18), 
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 18, colour = "black")) + 
  labs(title = "Distribution of CpGs significant at FDR = 0.05 across 12 blood cell types",
       x = "", y = "") +
  scale_fill_manual(values = cbPalette[2:3]) 
```

```{r}
ggplot(sign.pval.delta.mqtl.df, aes(x = Type, group = Treatment, fill = Treatment)) + 
    geom_density(aes(y = stat(count) / length(delta.meqtl.cpgs), color = Treatment), 
             stat = "count",
             alpha = 0.2) + 
    scale_y_continuous(labels = scales::percent) + 
    labs(x = " ", y = "", 
         title = "Distribution of CpGs significant at FDR = 0.05 across 12 blood cell types") + 
    theme(legend.title = element_blank(), 
          legend.position = "bottom", 
          panel.grid.major = element_blank(), 
          panel.background = element_blank(), 
          plot.title = element_text(size = 8), 
          axis.title = element_text(size = 8), 
          axis.text.x = element_text(angle = 0, hjust = 0.5)) +
    scale_color_manual(values = cbPalette[2:3]) +
    scale_fill_manual(values = cbPalette[2:3]) 
```
